RE: Issue #4 Add Comments
Comments have been added to the main HousingOutcomes.py script as well as some cleanup to make the pre-existing comments easier to read.

The spreadsheet will now have a default name and a default extension.

The Rundate class, which was previously an inactive element of the HousingOutcomes.py script.  This class has been moved into its own .py file so that it can be worked on idependently without making the code overly confusing.  I feel like having only a single class per .py file also conforms to best practices.
