+++ Housing Outcomes.py
@@ -3,71 +3,62 @@ __version__ = "1.12"
 
 """
 This script is for processing the Housing Services - Housing Outcomes v2.0 report that is used by
-the follow-ups specialistself.
+the follow-ups specialist.
 
 This script should identify when it is being run and adjust its date parameters to reflect being run
 on the first of the month or after it.  This well prevent future staff members from needing to
 fiddle with the code every time they run the report.
 """
 
+# import required libraries
 import pandas as pd
-from datetime import date
 from datetime import datetime
-from calendar import monthrange
-from dateutil.relativedelta import relativedelta
 from tkinter.filedialog import askopenfilename
 from tkinter.filedialog import asksaveasfilename
 
-class RunDate:
-    """
-    This class is not currently used by the CreateRequiredFollowUps class but I would like to
-    eventually make it so that the sheetnames output by that class are modified by the relation to
-    the current month.
-    """
-    def __init__(self):
-        self.today = datetime.now().date()
-        self.check_date()
-
-    def check_date(self):
-        if self.today.day <= 5:
-            last_month = self.today + relativedelta(months=-1)
-            end_of_month = date(
-                year=last_month.year,
-                month=last_month.month,
-                day=monthrange(last_month.year, last_month.month)[1]
-            )
-            return end_of_month
-        else:
-            end_of_month = date(
-                year=self.today.year,
-                month=self.today.month,
-                day=monthrange(self.today.year, self.today.month)[1]
-            )
-            return end_of_month
-
-
 class CreateRequiredFollowUps:
     def __init__(self, file_path):
+        # read the excel report into a pandas data frame
         self.raw_data = pd.read_excel(file_path)
+        # not currently in use
         self.run_date = RunDate()
+        # create a immutable list of unique months during which follow-ups are
+        # due
         self.month_range = set(
             [value.strftime("%B") for value in self.raw_data["Follow Up Due Date(2512)"]]
         )
+        # create month and year name variables for the name of the processed
+        # report
         self.current_month = datetime.now().month
+        self.current_year = datetime.now().year
 
     def process(self):
+        # create a local copy of the self.raw_data data frame
         data = self.raw_data
+        # initiate the ExcelWriter object variable
         writer = pd.ExcelWriter(
-            asksaveasfilename(title="Save the Required Follow-ups Report"),
+            asksaveasfilename(
+                title="Save the Required Follow-ups Report",
+                defaultextension=".xlsx",
+                initialfile="Required Follow-ups for {} {}".format(
+                    self.current_month,
+                    self.current_year
+                )
+            ),
             engine="xlsxwriter"
         )
+        # loop through the values of the self.month_range set creating dataframes
+        # where the value of Follow Up Due Date(2512) column is equal to the set
+        # item's value creating excel sheets for each of these data drames
         for month in self.month_range:
             month_data = data[
                 (data["Follow Up Due Date(2512)"].dt.strftime("%B") == month) &
                 data["Actual Follow Up Date(2518)"].isna()
             ].drop_duplicates(subset="Client Uid")
             month_data.to_excel(writer, sheet_name="{} Follow-Ups".format(month), index=False)
+        # create an excel sheet containing the raw data
         data.to_excel(writer, sheet_name="Raw Data", index=False)
+        # save the spreadsheet
         writer.save()
 
 if __name__ == "__main__":
